Extract filter to a separate file
